requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-core==2.14.1
starlette==0.27.0
anyio==3.7.1
httpx==0.25.2
httpcore==1.0.7
beautifulsoup4==4.12.2
soupsieve>=2.8
lxml==4.9.3
email-validator==2.1.0
dnspython==2.4.2
redis==5.0.1
async-timeout>=4.0.2
python-dotenv==1.0.0
slowapi==0.1.9
limits>=5.6.0
structlog==23.2.0
aiofiles==23.2.1
python-whois>=0.8.0
